reorganize and set up client class
