+++ src/resources/Bounty/constants.js
