+++ test/bounty.test.js
@@ -1,15 +1,17 @@
 import nock from 'nock'
 import fixtures from './fixtures'
-import { load } from '../src/bounty'
+import Bounties from '../src/bounties'
+
 
 nock.disableNetConnect()
 const mock = nock('https://api.bounties.network/')
+const bounties = new Bounties(1,2)
 
 describe('bounty module', () => {
     it('should load bounty with specified id', async () => {
         mock.get('/bounty/1/').reply(200, fixtures.bounty)
 
-        const bounty = await load(1)
+        const bounty = await bounties.bounty.load((1))
         expect(bounty).toEqual(fixtures.bounty)
     })
 
@@ -17,12 +19,12 @@ describe('bounty module', () => {
         const params = { platform__in: 'bounties' }
         mock.get('/bounty/1/').query(params).reply(200, fixtures.bounty)
 
-        const bounty = await load(1, params)
+        const bounty = await bounties.bounty.load(1, params)
         expect(bounty).toEqual(fixtures.bounty)
     })
 
     it('should fail if id does not exist', async () => {
         mock.get('/bounty/1234/').reply(400, fixtures.bounty);
-        expect(() => await load(1)).toThrow()
+        await expect(bounties.bounty.load(1)).rejects.toThrow(new Error())
     })
 })
